Merge Overlapping Variants
When calling variants, the HaplotypeCaller only weighs evidence
between variants that start at the exact same position.  This creates
problems when deletions or insertions overlap, because than the sequences
that have the overlapping insertion or deletion are counted as "reference"
despite not having the reference sequence.  This change tries to fix this
by:

1 - Merging all overlapping variants into one "Super variant" with the
    same start position for each.
2 - Unmerging the variants at the end of the calling step so that
    the highest start position is used.